Centralize operations on hardcoded paths
refs #54
